+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 7.1
  - Added default choice for some yes/no question. Regexp improved
  - Tested compatibility with Backbox 4.7 and Parrot 3.6
- - Improved curl method. Now autodetect if http proxy is set and needed
+ - Improved curl and wget methods. Now autodetect if http proxy is set and needed
  
 ### 7.02
  - Improved internet checking. Now ip address and host based methods