+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Tested compatibility with Backbox 4.7 and Parrot 3.6
  - Improved curl and wget methods. Now autodetect if http proxy is set and needed
  - WPS Reaver attacks modified. Removed --auto (-a) argument causing problems on some Reaver versions
+ - Dockerfile updated
 
 ### 7.02
  - Improved internet checking. Now ip address and host based methods